This pull request optimizes the fetching of images for the tracklabels plugin. 

It leverages web browser cookies to reduce web server load and increase page loading speeds. It checks with your web browser to see if a request for a special image has failed within the last 30 days. If so, it won't ask again unless you clear cookies.

It also uses `sendFile()` in `~/php/util.php` to cache the generic image sent, when a label or tracker doesn't have a special image. This is significantly faster for page refreshes than performing a 302 URL redirection to the location of the generic image.

When no special image is available, these requests now complete in ping + 30ms for a fresh load or 60ms for a cached load. Before this patch, there were instances of these requests taking 500ms to 800ms to complete.