@Novik 2021-12-27
1) https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Set-Cookie

>A cookie-name can contain any US-ASCII characters except for: the control character, space, or a tab. It also must not contain a separator characters like the following: ( ) < > @ , ; : \ " / [ ] ? = { }.

"Any US-ASCII characters". But label may contain non-US-ASCII characters.

2) For case (!isset($_REQUEST["label"] && isset($_REQUEST["tracker"]) && isset($_COOKIE[$_REQUEST["tracker"]]) we always will have /trackers/unknown.png as result. Why?