@stickz 2021-12-28
1. I plan on improving the cookies to work for separator characters by encoding the cookie name in url format. I will also add a regex expression to detect if the label is not in US-ASCII format. This performance improvement should still benefit most people who use ruTorrent. I don't think it's worth discarding this change for a handful of other use cases.

2. The way issets are used here is very confusing. They do work as intended and will not always return `unknown.png`. I will add a few code comments, when I submit anther commit to this pull request. These will help to clarify how the issets code is used.