@Novik 2021-12-29
> I plan on improving the cookies to work for separator characters by encoding the cookie name in url format. 

We already have encoded value in the $_REQUEST["label"]. But for unknown reason you construct cookie name from it after decoding.

> They do work as intended and will not always return unknown.png

I can't see how this may be... BTW, what we will have for case isset($_REQUEST["label"]) && isset($_COOKIE[$cookie])? As I may see, also unknown.png.
Sorry, but I can't understand, how this may work.