@stickz 2022-01-02
I'm closing this pull request because it's not compatible with #2247. I will open a new pull request to optimize the track labels plugin after #2247 is dealt with. It will not include cookies. Just the image caching instead of 302 redirection improvement.

Refactoring the PHP code significantly reduced the amount of time required to complete this task. The code refactor is a better solution than cookies. It dropped the time required to complete this task down to the approximate network transfer time.